Add manual sync time editing for mail accounts
Introduces EditSyncTime actions to MailAccountsController, a new EditSyncTimeViewModel, and a corresponding view to allow manual adjustment of the LastSync timestamp for mail accounts. Updates resource files for localization and adds an 'Edit Sync Time' button to the account details view. Also refactors folder retrieval to use GraphEmailService for M365 accounts.